Removed unuses code bits, extracted hashForAnon to a utils file
